CR: Pull options menu into separate component
