+++ packages/client/components/TeamPrompt/TeamPromptOptionsMenu.tsx
@@ -3,32 +3,16 @@ import graphql from 'babel-plugin-relay/macro'
 import React from 'react'
 import {useFragment} from 'react-relay'
 import useAtmosphere from '~/hooks/useAtmosphere'
-import {MenuPosition} from '~/hooks/useCoords'
-import useMenu from '~/hooks/useMenu'
+import {MenuProps} from '~/hooks/useMenu'
 import useMutationProps from '~/hooks/useMutationProps'
 import EndTeamPromptMutation from '~/mutations/EndTeamPromptMutation'
 import {ICON_SIZE} from '~/styles/typographyV2'
 import {TeamPromptOptionsMenu_meeting$key} from '~/__generated__/TeamPromptOptionsMenu_meeting.graphql'
 import {PALETTE} from '../../styles/paletteV3'
-import CardButton from '../CardButton'
 import Icon from '../Icon'
-import IconLabel from '../IconLabel'
 import Menu from '../Menu'
 import MenuItem from '../MenuItem'
 import {MenuItemLabelStyle} from '../MenuItemLabel'
-
-const Options = styled(CardButton)({
-  position: 'absolute',
-  top: 0,
-  right: 0,
-  color: PALETTE.SLATE_700,
-  height: 32,
-  width: 32,
-  opacity: 1,
-  ':hover': {
-    backgroundColor: PALETTE.SLATE_300
-  }
-})
 
 const StyledIcon = styled(Icon)({
   color: PALETTE.SLATE_600,
@@ -43,12 +27,11 @@ const OptionMenuItem = styled('div')({
 
 interface Props {
   meetingRef: TeamPromptOptionsMenu_meeting$key
+  menuProps: MenuProps
 }
 
 const TeamPromptOptionsMenu = (props: Props) => {
-  const {togglePortal, originRef, menuPortal, menuProps} = useMenu(MenuPosition.UPPER_RIGHT)
-
-  const {meetingRef} = props
+  const {meetingRef, menuProps} = props
 
   const meeting = useFragment(
     graphql`
@@ -64,7 +47,7 @@ const TeamPromptOptionsMenu = (props: Props) => {
   const atmosphere = useAtmosphere()
   const {onCompleted, onError} = useMutationProps()
 
-  const renderedMenu = (
+  return (
     <Menu ariaLabel={'Edit the meeting'} {...menuProps}>
       <MenuItem
         key='copy'
@@ -81,15 +64,6 @@ const TeamPromptOptionsMenu = (props: Props) => {
         }}
       />
     </Menu>
-  )
-
-  return (
-    <>
-      <Options ref={originRef} onClick={togglePortal}>
-        <IconLabel ref={originRef} icon='more_vert' />
-      </Options>
-      {menuPortal(renderedMenu)}
-    </>
   )
 }
